Divide the header into parts for readablity
